added timing section to sudoku_tester for manual run time testing.
sudoku_solver.h: Fixed logic errors that were causeing longer than 
necessary run time by making adding an if statement to check if a guess 
was clearly not viable before creating a new puzzle object to hold it.  
There was also logic errors in exiting Solver() upon finding a solution 
the solution was saved but searching continued until many extra 
possibilities were atempted and a final fail caused the search to end 
but with the correct answer already passed back by reference. run time 
of Solver() was reduced from 26ms to 8ms with the if statment then from 
8ms to 5ms with the exit fix.